allows to choose encoding for a key or a value
previous code was quite naive and was just trying to encode the value as a
binary term when needed. It's now possible to choose the key encoding or value
encoding when creating a storage engine.

Different encoding are:
- raw -> no encoding
- term -> term_to_binary is used
- {term, Opts} => like above but with options
- sext -> use the library sext from uwiger to encode terms

Note: the ets backend doesn't encode the value at all since it's already
handed by the backend.